Add functionality for specifying key age threshold per SA
This removes the ability for users to specify multiple key sources for a
single SA, which was a bad idea in the first place. Users can now just
set all their sources in a single keySource block, which should be fine.
This had to be done, in order for there the SA<-->key-age-threshold to
be 1:1