Document the Base64Decode circleci parameter (#508)
It was absolutely necessary for our particular use-case (which is
hosting the runner images in GCR), but I encountered it only because I
went into the code to see how easy it would be to change a couple of
things we needed. This was one of them, and the most important, so it
was a pleasant surprise to find it!